pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mini_latent_pd"
version = "0.0.1"
description = "A pytorch package for doing protein dynamics in latent spaces defined by pretrained protein folding models"
authors = [
  { name = "Moust Holmes", email = "moust.holmes@bio.ku.dk" },
]
license = { file = "LICENSE" }
keywords = ["machine learning", "flow matching", "generative models"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "invoke>=2.2.0",
    "hydra-core>=1.3.2",
    "pydantic>=2.0.0",
    "loguru>=0.7.0",
    "torch>=2.0.0",
    "lightning>=2.2.0",
    "omegaconf>=2.3.0",
    "wandb>=0.16.0",
    "torchvision>=0.15.0",
    "matplotlib>=3.2.0",
    "scikit-learn",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=23.9.1",
    "ruff>=0.2.2",
    "pre-commit>=3.5.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
    "invoke>=2.2.0",
    "mypy>=1.8.0",
    "nbqa>=1.7.1",
    "nbstripout>=0.7.1",
    "bandit[toml]>=1.7.7",
    "pytest-cov>=4.1.0",
    "types-PyYAML>=6.0.12",
    "types-requests>=2.31.0",
    "types-setuptools>=69.0.0",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["I", "N", "E", "W"]
ignore = ["N812", "N813"]  # Ignore rules about import casing conventions

[tool.coverage.run]
omit = ["tests/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = [
    "--strict-markers",
    "--doctest-modules",
    "--color=yes",
    "--import-mode=importlib",
]
pythonpath = ["src"]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]
